+++ src/test/java/com/jnape/palatable/lambda/functions/specialized/BiPredicateTest.java
@@ -2,7 +2,9 @@ package com.jnape.palatable.lambda.functions.specialized;
 
 import org.junit.Test;
 
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 public class BiPredicateTest {
@@ -49,5 +51,11 @@ public class BiPredicateTest {
         assertFalse(equals.test("b", "a"));
         assertFalse(equals.negate().test("a", "a"));
         assertTrue(equals.negate().test("b", "a"));
+    }
+
+    @Test
+    public void flip() {
+        BiPredicate<String, String> equals = String::equals;
+        assertThat(equals.flip(), instanceOf(BiPredicate.class));
     }
 }